refactor(constant): 상수 이름 수정
Constant -> ConstantNumbers
정의된 정수들의 모음이라는 것을 이름으로 전달하기 위해 변경합니다.

BILL -> RECEIPT
영수증임을 표현하기위해 변경합니다.
