+++ src/main/java/store/constant/ConstantNumbers.java
@@ -1,8 +1,8 @@
 package store.constant;
 
-public class Constant {
-    public final static int MIN_STOCK_PRICE = 100;
-    public final static int MAX_STOCK_PRICE = 100_000;
+public class ConstantNumbers {
+    public static final int MIN_STOCK_PRICE = 100;
+    public static final int MAX_STOCK_PRICE = 100_000;
     public static final int MIN_LENGTH_ITEM_NAME = 2;
     public static final int MAX_LENGTH_ITEM_NAME = 10;
     public static final int MIN_LENGTH_PROMOTION_NAME = 4;
