+++ src/main/java/store/constant/Errors.java
@@ -1,14 +1,14 @@
 package store.constant;
 
-import static store.constant.Constant.EXACT_GET_PROMOTION;
-import static store.constant.Constant.MAX_BUY_PROMOTION;
-import static store.constant.Constant.MAX_LENGTH_ITEM_NAME;
-import static store.constant.Constant.MAX_LENGTH_PROMOTION_NAME;
-import static store.constant.Constant.MAX_STOCK_PRICE;
-import static store.constant.Constant.MIN_BUY_PROMOTION;
-import static store.constant.Constant.MIN_LENGTH_ITEM_NAME;
-import static store.constant.Constant.MIN_LENGTH_PROMOTION_NAME;
-import static store.constant.Constant.MIN_STOCK_PRICE;
+import static store.constant.ConstantNumbers.EXACT_GET_PROMOTION;
+import static store.constant.ConstantNumbers.MAX_BUY_PROMOTION;
+import static store.constant.ConstantNumbers.MAX_LENGTH_ITEM_NAME;
+import static store.constant.ConstantNumbers.MAX_LENGTH_PROMOTION_NAME;
+import static store.constant.ConstantNumbers.MAX_STOCK_PRICE;
+import static store.constant.ConstantNumbers.MIN_BUY_PROMOTION;
+import static store.constant.ConstantNumbers.MIN_LENGTH_ITEM_NAME;
+import static store.constant.ConstantNumbers.MIN_LENGTH_PROMOTION_NAME;
+import static store.constant.ConstantNumbers.MIN_STOCK_PRICE;
 
 public enum Errors {
