+++ src/main/java/store/presentation/view/Message.java
@@ -12,16 +12,16 @@ public enum Message {
     INPUT_MEMBERSHIP("멤버십 할인을 받으시겠습니까? (Y/N)"),
     INPUT_BUY_MORE("감사합니다. 구매하고 싶은 다른 상품이 있나요? (Y/N)"),
 
-    BILL_HEADER("===========W 편의점============="),
-    BILL_PRODUCTS_HEADER("상품명\t\t수량\t금액"),
-    BILL_PRODUCT("%s \t\t%d \t%,d\n"),
-    BILL_PROMOTION_HEADER("===========증\t정============="),
-    BILL_PROMOTION("%s\t\t%d"),
-    BILL_SEPARATOR("=============================="),
-    BILL_TOTAL_PRICE("총구매액\t\t%d\t%,d"),
-    BILL_PROMOTION_DISCOUNT("행사할인\t\t\t-%,d"),
-    BILL_MEMBERSHIP_DISCOUNT("멤버십할인\t\t\t-%,d"),
-    BILL_PAY("내실돈\t\t\t %,d");
+    RECEIPT_HEADER("===========W 편의점============="),
+    RECEIPT_PRODUCTS_HEADER("상품명\t\t수량\t금액"),
+    RECEIPT_PRODUCT("%s \t\t%d \t%,d\n"),
+    RECEIPT_PROMOTION_HEADER("===========증\t정============="),
+    RECEIPT_PROMOTION("%s\t\t%d"),
+    RECEIPT_SEPARATOR("=============================="),
+    RECEIPT_TOTAL_PRICE("총구매액\t\t%d\t%,d"),
+    RECEIPT_PROMOTION_DISCOUNT("행사할인\t\t\t-%,d"),
+    RECEIPT_MEMBERSHIP_DISCOUNT("멤버십할인\t\t\t-%,d"),
+    RECEIPT_PAY("내실돈\t\t\t %,d");
 
     private final String message;
